Add WeightedPredictor ABC, rework learning module
