+++ VSharp.ML.AIAgent/ml/model_wrappers/protocols.py
@@ -1,16 +1,21 @@
-from abc import abstractmethod
-from typing import Protocol
+from abc import ABC, abstractmethod
 
 from common.game import GameState
 
 
-class Named(Protocol):
+class Named(ABC):
     @abstractmethod
     def name(self) -> str:
         raise NotImplementedError
 
 
-class Predictor(Named, Protocol):
+class Predictor(Named, ABC):
     @abstractmethod
     def predict(self, input: GameState):
+        raise NotImplementedError
+
+
+class WeightedPredictor(Predictor, ABC):
+    @abstractmethod
+    def weights(self) -> list:
         raise NotImplementedError
